+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Increment the:
 
 * [DEPRECATION] Deprecate ZPAGES
   [#2291](https://github.com/open-telemetry/opentelemetry-cpp/pull/2291)
+* [EXPORTER] Remove explicit timestamps from metric points exported by Prometheus
+  [#2324](https://github.com/open-telemetry/opentelemetry-cpp/pull/2324)
+* [EXPORTER] Handle attribute key collisions caused by sanitation
+  [#2324](https://github.com/open-telemetry/opentelemetry-cpp/pull/2326)
 
 ## [1.11.0] 2023-08-21